shanyraq2/requirements.txt
Django==4.2
djangorestframework==3.14
djangorestframework-simplejwt>=5.3.1
django-cors-headers==4.3
Pillow==10.0
python-decouple==3.8
setuptools>=70
